+++ package.json
@@ -36,9 +36,7 @@
     "version": "auto-changelog -p && git add CHANGELOG.md",
     "cypress:lin": "export NODE_ENV=development && npx cypress open",
     "cypress:win": "set \"NODE_ENV=development\" && npx cypress open",
-    "postinstall": "patch-package",
-    "prepare-native": "npm run nodeify",
-    "nodeify": "./node_modules/.bin/rn-nodeify --hack --install"
+    "postinstall": "patch-package"
   },
   "repository": {
     "type": "git",
@@ -97,7 +95,7 @@
     "mnid": "^0.1.3",
     "mobile-device-detect": "^0.2.3",
     "moment": "^2.24.0",
-    "node-libs-browser": "^2.2.1",
+    "node-libs-react-native": "^1.0.3",
     "numeral": "^2.0.6",
     "patch-package": "^6.2.0",
     "path-browserify": "0.0.0",
@@ -204,7 +202,6 @@
     "react-hot-loader": "^4.6.5",
     "react-native-dotenv": "^0.2.0",
     "react-test-renderer": "16.9.0",
-    "rn-nodeify": "tradle/rn-nodeify",
     "serve": "^11.0.0",
     "source-map-explorer": "^2.2.2",
     "webpack-bundle-analyzer": "^3.6.0"
@@ -234,57 +231,5 @@
   ],
   "auto-changelog": {
     "tagPattern": "^v[\\.0-9]+$"
-  },
-  "react-native": {
-    "zlib": "browserify-zlib",
-    "console": "console-browserify",
-    "constants": "constants-browserify",
-    "crypto": "react-native-crypto",
-    "dns": "dns.js",
-    "net": "react-native-tcp",
-    "domain": "domain-browser",
-    "http": "@tradle/react-native-http",
-    "https": "https-browserify",
-    "os": "react-native-os",
-    "path": "path-browserify",
-    "querystring": "querystring-es3",
-    "fs": "react-native-level-fs",
-    "_stream_transform": "readable-stream/transform",
-    "_stream_readable": "readable-stream/readable",
-    "_stream_writable": "readable-stream/writable",
-    "_stream_duplex": "readable-stream/duplex",
-    "_stream_passthrough": "readable-stream/passthrough",
-    "dgram": "react-native-udp",
-    "stream": "stream-browserify",
-    "timers": "timers-browserify",
-    "tty": "tty-browserify",
-    "vm": "vm-browserify",
-    "tls": false
-  },
-  "browser": {
-    "zlib": "browserify-zlib",
-    "console": "console-browserify",
-    "constants": "constants-browserify",
-    "crypto": "react-native-crypto",
-    "dns": "dns.js",
-    "net": "react-native-tcp",
-    "domain": "domain-browser",
-    "http": "@tradle/react-native-http",
-    "https": "https-browserify",
-    "os": "react-native-os",
-    "path": "path-browserify",
-    "querystring": "querystring-es3",
-    "fs": "react-native-level-fs",
-    "_stream_transform": "readable-stream/transform",
-    "_stream_readable": "readable-stream/readable",
-    "_stream_writable": "readable-stream/writable",
-    "_stream_duplex": "readable-stream/duplex",
-    "_stream_passthrough": "readable-stream/passthrough",
-    "dgram": "react-native-udp",
-    "stream": "stream-browserify",
-    "timers": "timers-browserify",
-    "tty": "tty-browserify",
-    "vm": "vm-browserify",
-    "tls": false
   }
 }
