--- patches/add-v8-jit.patch
@@ -1,59 +1,0 @@
-From 5f05997c7fea6f6f8572af2c0993fa94edfc848f Mon Sep 17 00:00:00 2001
-From: Hadar Rottenberg <hadar@gooddollar.org>
-Date: Sun, 12 Jan 2020 23:37:47 +0200
-Subject: [PATCH] add: v8 (JIT missing)
-
----
- android/app/build.gradle |  7 ++++++-
- android/build.gradle     |  9 +++++++--
- 4 files changed, 27 insertions(+), 3 deletions(-)
-
-diff --git a/android/app/build.gradle b/android/app/build.gradle
-index d057dbda..a1d508d4 100644
---- a/android/app/build.gradle
-+++ b/android/app/build.gradle
-@@ -177,13 +177,18 @@ android {
- 
-         }
-     }
-+    packagingOptions {
-+        // Make sure libjsc.so does not packed in APK
-+        exclude "**/libjsc.so"
-+    }
- }
- 
- dependencies {
-     implementation project(':react-native-vector-icons')
-     implementation fileTree(dir: "libs", include: ["*.jar"])
-     implementation "com.facebook.react:react-native:+"  // From node_modules
--
-+    // Add v8-android - prebuilt libv8android.so into APK 
-+    implementation 'org.chromium:v8-android:+'
-     implementation "androidx.annotation:annotation:1.1.0"
- 
-     if (enableHermes) {
-diff --git a/android/build.gradle b/android/build.gradle
-index 28f7ec64..b2faac51 100644
---- a/android/build.gradle
-+++ b/android/build.gradle
-@@ -24,11 +24,16 @@ allprojects {
-         mavenLocal()
-         maven {
-             // All of React Native (JS, Obj-C sources, Android binaries) is installed from npm
--            url("$rootDir/../node_modules/react-native/android")
-+            // url("$rootDir/../node_modules/react-native/android")
-+            // Replace AAR from original RN with AAR from react-native-v8
-+            url("$rootDir/../node_modules/react-native-v8/dist")
-         }
-         maven {
-+            // prebuilt libv8android.so
-+            url("$rootDir/../node_modules/v8-android-jit-nointl/dist")
-             // Android JSC is installed from npm
--            url("$rootDir/../node_modules/jsc-android/dist")
-+            //url("$rootDir/../node_modules/jsc-android/dist")
-+
-         }
- 
-         google()
--- 
-2.21.0
